+++ src/test/resources/test-projects/helper-script/app/counter.py
@@ -1,2 +1,0 @@
-def plusone(_number):
-    return _number + 1