+++ src/test/resources/test-projects/whole-pipeline/app/test_counter.py
@@ -1,6 +1,0 @@
-import unittest
-import counter
-
-class TestCounter(unittest.TestCase):
-    def test_plusone(self):
-        self.assertEqual(2, counter.plusone(1))