+++ CHANGELOG.md
@@ -7,12 +7,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Unreleased
 
-* Make the resource path lookup customizable for `JenkinsPipelineSpecification#loadPipelineScriptForTest`. To define your own path set `JenkinsPipelineSpecification.scriptClassPath`.
-
 ### Fixed
 
 * Added `<repositories>` section to all poms pointing to the Jenkins Release repository - so that Jenkins artifacts can be successfully located during builds.
-* Emit more-specific exception when tests call a Groovy method with an incorrect signature
+
+### Updated
+	
+* Emit more-specific exception when tests call a Groovy method with an incorrect signature.
+
+### Added
+
+* Make the resource path lookup customizable for `JenkinsPipelineSpecification#loadPipelineScriptForTest`. To define your own path set `JenkinsPipelineSpecification.scriptClassPath`.
 
 ## 2.0.1