+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Updated
 	
 * Emit more-specific exception when tests call a Groovy method with an incorrect signature.
+* Clarified documentation for using "global" variables in tests.
 
 ### Added